Добавит два презентера для фильма в списке и попапа
